style: Restore password entry width harder
Commit ee0538807 restored the fixed password entry width, but only
for entries nested under .prompt-dialog-password-layout. That
does not include the network secret dialog, so go back to the
original styling prior to commit 8d065b558e where the width
is set for all password entries, and the mount operation overrides
it with a more specific rule.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3514>