status/volume: Show OSD volume slider on plugging unplugging headphones
OSD Volume popup appears when a user connects or disconnects
earphones/headphones/headsets.

Shows and warns the user that the volume may be set to high
or low when headphones are connected or disconnected.

This is a nice bit of feedback to the user that something has
changed, and something that other platforms do.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7931
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3531>