loginDialog: Add Switch User button in corner
The latest mockups for login/unlock here:

https://gitlab.gnome.org/Teams/Design/os-mockups/-/blob/master/lock-login/login-unlock.png

seek to harmonize login and unlock by changing the cancel button to a
Switch User... button.

This commit does that except at the Not Listed... username prompt,
because it would be confusing to have a Switch User... button at a
screen asking for a username.