data: Really strip .desktop suffix from folder directory
Commit 730fcf34e9 was meant to strip the .desktop suffix from the
desktop directory file, but failed to actually do so: The replace()
method is called on the string parameter before it is passed to
the meson command which substitutes `@BASENAME@`, so it doesn't
have any effect at all.

Really fix the issue by providing the full output name.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3567>