ci: Change `update-mutter` to also install to sysext dir
Out development toolbox contains an `update-mutter` command to
make it easier to keep mutter up-to-date for people who only
want to deal with gnome-shell.

To allow them to make use of the new sysext tooling, change the
command to install both to the toolbox root and the sysext
directory.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3498>