main: Disable text mipmapping with env variable
In shell_fonts_init we disable text mipmapping by setting a property on the
font map retrieved by clutter_get_font_map.
clutter_get_font_map will initialize the font map in a thread, which means
we might have a thread calling getenv while mutter is setting up the display
environment variables. setenv/getenv are not thread safe [1] so we may crash.

Instead, we can use CLUTTER_DISABLE_MIPMAPPED_TEXT to disable mipmapping, which
will take effect later, whenever the font map is initialized.
Ironically, that is also a setenv so move it to before any threads are started.

This is a stopgap for the most frequent crashes as seen in Ubuntu and Fedora
crash collectors, while a complete fix would be to set the display environment
variables earlier, or not set them at all in the gnome-shell process [2].

In GNOME 47 the early initialization of the font map was removed with [3],
which is why we don't see crashes collected on those versions.
This has the same effect, but is more easily backportable as a self-contained
change.

[1]: improvements are on the way for glibc, https://sourceware.org/git/?p=glibc.git;a=commit;h=7a61e7f557a97ab597d6fca5e2d1f13f65685c61
[2]: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2888
[3]: https://gitlab.gnome.org/GNOME/mutter/-/commit/60edafe3ebe56b46f561599f7602c3867fd7729a,
     https://gitlab.gnome.org/GNOME/gnome-shell/-/commit/ccc46d3f5c1d7b6b9c7e04b825982df05b0e60dc

Related: https://gitlab.gnome.org/GNOME/mutter/-/issues/2463
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3571>