gdmUtil: Nullify hold when releasing it
A "hold" is an object for inhibiting a Batch from
completing. A Batch is a pre-Promise era object for
doing a series of operations asynchronously.

At the moment, we can end up in a situation where we
try to release a hold more than once. This create a
harmless error in the log.

Eventually we should scrap all Batch usage altogether,
but for now this commit nullifies the hold when it is
released to silence the log spew.