gdm: Add support for Smartcard in authServicesSwitchable
This allows selecting smartcard as a login method.

Currently smartCard mechanism will be displayed for all users,
independently of whether they have a smartcard cert mapped or not.
In the future, if an empty smartcard mechanism is returned only for users
with a cert mapped, we can hide the smartcard mechanism for users without
a cert.

When a smartcard is inserted the switchable service will be restarted to
check if a cert is available for the current user. This won't change the
authentication state (if it was doing password auth, that won't change).

https://gitlab.gnome.org/Teams/Design/os-mockups/-/blob/master/lock-login/login-unlock.png