gdm: Add support for Smartcard in authServicesSSSDSwitchable
This allows selecting smartcard as a login method.

While SSSD can inform for each user what mechanisms are available, it
can't know if a user has smartcard mechanism available until a valid smartcard
is inserted. To allow smartcard mechanism, if it's enabled in GDM, add an empty
smartcard mechanism for every user.

When a smartcard is inserted, authServicesSSSDSwitchable will be restarted to
check if a cert is available for the current user. This won't change the
authentication state (if it was doing password auth, that won't change).

To avoid multiple events, SmartcardManager won't connect to tokens with
'/login_token', those are aliases for already connected tokens.