gdm: Add support for Smartcard in authServicesSSSDSwitchable
This allows selecting smartcard as a login method.

While SSSD can inform for each user what mechanisms are available, it
can't know if a user has smartcard mechanism available until a valid smartcard
is inserted.

When a smartcard is inserted, authServicesSSSDSwitchable will be restarted to
check if a cert is available for the current user. This won't change the
authentication state (if it was doing password auth, that won't change).