loginDialog: When updating disable user list ask for username
Right now if the disable-user-list key is changed we do
a reset that asks for a username via PAM instead of via
gnome-shell. This is different than Not Listed? behavior
and can get in the way off smartcard handling, so change
it to explicitly ask for the username up front.