Add asyncMux to avoid race conditions in async functions
When several async functions share access to common elements,
it's possible to have race condition problems. For example,
when an extension is enabled, it is deleted from the list of
disabled extensions and added to the list of enabled extensions.
This can cause a race condition because both cases call the
same asynchronous function, resulting in the creation of two
instances of each extension, and the `enable` method of each
one being called twice. This is possible because the method
`onEnabledExtensionsChanged` can be called both from a signal
of a change in the `disabled-extensions` property, or from
a signal of a change in the `enabled-extensions` property.
Since this method is async and does several `await` calls,
two very close changes to those properties can trigger it
the second time when the first time is in the middle of an
`await`, and, thus, in an incomplete state.

This patch defines an `AsyncMutex` object that allows to
ensure that, if an async function is called again while
already processing a call, won't execute until any previous
call has ended and completed their operations. This is, it works
like a mutex semaphore, but designed for Javascript async
functions/methods. GLib mutexes can't be used in this case
because they are tied to the C language and presume threading/
multiprocess, something that doesn't exist in javascript's
asynchronous programming.

To work with it, an `AsyncMutex` object is created for each
critical section. Then, the `hold()` method must be awaited when
entering the critical section, and the `release()` method must be
called just after exiting it. The first worker will run as usual,
but if it does an `await` inside the critical section and a new
worker tries to enter it, the `await asyncMutex.hold()` call
will wait until the first worker calls `asyncMutex.release()`.
The `AsyncMutex` class implements a waiting queue, thus if
several workers are awaiting, when the worker that owns the
mutex releases it, the ownership will be passed to the first
worker that waited when doing the `hold()` call. Each worker
will receive ownership of the mutex in the same order than they
requested it when calling the `hold()` method.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3640>