webLogin: Move web login buttons management to the mechanism
The behavior of WebUI may change depending on the protocol
implementation, so we need to add some flexibility to it so that it's
not tied to the protocol in use.

For example in other protocols we may not need to show any "Done"
button, but scanning results are just reported promptly.

While other buttons can be added (for example to trigger a URI refresh).

So move this at lower level, and allow handling this differently based
on the protocol