runDialog: Fix launching shell command in terminal
At one point Clutter became stricter when handling keybindings,
and the text's `activate` signal is no longer emitted when
additional modifiers are pressed.

To fix launching a command in the terminal when activating
with ctrl, make the entry non-activatable and move activation
to a `key-press-event` handler.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7516
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3683>