status/doNotDisturb: Add new quick toggle
It is a simple on/off control, which makes it a good fit for
quick settings. It also makes sense to expose the feature together
with other system controls, and while this somewhat removes the
control from the context of notifications, that doesn't seem to
be an issue on other platforms.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8339