dateMenu: Stop indicating do-not-disturb status
The control for enabling/disabling do-not-disturb will be moved into
quick settings. It makes sense to then also move the indicator, so stop
showing the do-not-disturb status in the top bar's messages indicator.

Instead, just hide the indicator altogether when do-not-disturb
is active.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3691>