breakManager: Use uninhibitable idle watch
If we use a regular idle watch, any session inhibitors can prevent the
watch from triggering, leading to the break manager to never consider
the break as having been completed until the inhibitor is gone. This
could lead to situations in which the system appears to be stuck with
a dimmed screen.

Also start the idle time when the idle watch is created, otherwise any
previous idle time would count towards it, causing the watch to trigger
early, leaving it inconsistent with the rest of the break manager state.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8378
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3709>