appDisplay: Clicking outside folders does not consistently close them
This is because the container of the dialog changes sizes after a
refresh, for reasons that are not fully explained. A better condition
is to check explicitly that the click doesn't fall into the dialog.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7541
Co-authored-by: Florian Müllner <fmuellner@gnome.org>
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3713>
(cherry picked from commit 5de29efbe5a3e08388984586f906a396a5ea5a13)