status/brightness: Port to new MetaBacklight API
Mutter has taken over controlling the backlights of the system from
gnome-settings-daemon. It now exposes MetaMonitors which can be reached
from MetaMonitorManager, and any of the MetaMonitors can have a
controllable backlight.

For now, we keep the current UX and simply set the backlight percentage
of the primary monitor's backlight (or the first monitor's backlight, if
the primary monitor doesn't have a backlight) on all monitors.

g-s-d is still responsible for reacting to keyboard shortcuts and
popping up the OSD notification, but that might change in the future.

The UI may change to make it possible to control individual backlights
in the future.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3715>