network: Don't assume captive portal indicates wi-fi
Turns out my AT&T Internet gateway acts as a captive portal for about
two minutes after it is started. It's wrong to assume captive portal
indicates Wi-Fi.

Also, change the terminology from "Sign Into" to "Sign In to" to be
pedantic. "Sign Into" seems to be pretty widely used on the web, so
maybe users won't notice, but "Into" is supposed to be a preposition,
not a verb.

Closes https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8330