fileUtils: Handle symlinks when moving/deleting dirs
We can handle them like regular files for the operations in question,
but we have to do so explicitly. Otherwise the link will not be
moved or removed, and in the latter case prevent the parent
directory from being removed as well (as it ends up non-empty).

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8417
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3721>