test: Don't append test env variables
If the specified variable already exists in the environment, append()
will append the value instead of overriding it. This isn't what we
want for the existing use case ("TZ=Europe/Madrid:UTC"), nor for
the case we are about to add.

Change the setup to use set() instead.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3721>